requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
scikit-image==0.21.0
Pillow==10.1.0
numpy==1.24.3
scipy==1.11.3
matplotlib==3.8.0
plotly==5.17.0
rembg==2.0.50
